backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
google-generativeai==0.7.2
firebase-admin==6.5.0
python-dotenv==1.0.1
pydantic==2.7.0
httpx==0.27.0
